Small crosschain swaps cleanup (#4012)
* small cleanup

* added missing file
